requirements.txt
warp-lang==1.1.0
trimesh
pyyaml
numpy==1.26.4
scipy
transformers
scikit-learn
scikit-image
pyvista
opencv-contrib-python
open3d
moviepy
matplotlib
kornia
ipdb
h5py
gradio
ffmpeg
einops
dill
av
timm
omegaconf
hydra-core
wandb